+++ uvicorn/lifespan/on.py
@@ -1,11 +1,15 @@
 import asyncio
 import logging
+from asyncio import Queue
+
+from uvicorn import Config
+from uvicorn._types import LifespanReceiveMessage, LifespanScope, LifespanSendMessage
 
 STATE_TRANSITION_ERROR = "Got invalid state transition on lifespan protocol."
 
 
 class LifespanOn:
-    def __init__(self, config):
+    def __init__(self, config: Config) -> None:
         if not config.loaded:
             config.load()
 
@@ -13,12 +17,12 @@ class LifespanOn:
         self.logger = logging.getLogger("uvicorn.error")
         self.startup_event = asyncio.Event()
         self.shutdown_event = asyncio.Event()
-        self.receive_queue = asyncio.Queue()
+        self.receive_queue: "Queue[LifespanReceiveMessage]" = asyncio.Queue()
         self.error_occured = False
         self.startup_failed = False
         self.should_exit = False
 
-    async def startup(self):
+    async def startup(self) -> None:
         self.logger.info("Waiting for application startup.")
 
         loop = asyncio.get_event_loop()
@@ -33,7 +37,7 @@ class LifespanOn:
         else:
             self.logger.info("Application startup complete.")
 
-    async def shutdown(self):
+    async def shutdown(self) -> None:
         if self.error_occured:
             return
         self.logger.info("Waiting for application shutdown.")
@@ -41,10 +45,13 @@ class LifespanOn:
         await self.shutdown_event.wait()
         self.logger.info("Application shutdown complete.")
 
-    async def main(self):
+    async def main(self) -> None:
         try:
             app = self.config.loaded_app
-            scope = {"type": "lifespan"}
+            scope: LifespanScope = {
+                "type": "lifespan",
+                "asgi": {"version": self.config.asgi_version, "spec_version": "2.0"},
+            }
             await app(scope, self.receive, self.send)
         except BaseException as exc:
             self.asgi = None
@@ -61,7 +68,7 @@ class LifespanOn:
             self.startup_event.set()
             self.shutdown_event.set()
 
-    async def send(self, message):
+    async def send(self, message: LifespanSendMessage) -> None:
         assert message["type"] in (
             "lifespan.startup.complete",
             "lifespan.startup.failed",
@@ -86,5 +93,5 @@ class LifespanOn:
             assert not self.shutdown_event.is_set(), STATE_TRANSITION_ERROR
             self.shutdown_event.set()
 
-    async def receive(self):
+    async def receive(self) -> LifespanReceiveMessage:
         return await self.receive_queue.get()